Adding numerical dice total to roll results
Getting the string results from a dice roll are fine when you're simply
outputting the result. However, to do any "real work" with the results,
I need the value of the roll in a numerical fashion as well. I've added
a new struct so that calling methods can now choose which to use, and
the struct can still be printed as a string like the old result. Now I
can roll a dice and use the result as a number, which is much easier to
work with programatically.